+++ stop_words/tests.py
@@ -1,15 +1,18 @@
 """
 Tests for stop-words
 """
+import random
 from unittest import TestCase
 from unittest import TestSuite
 from unittest import TestLoader
 
 import stop_words
-from stop_words import StopWordError
 from stop_words import get_stop_words
 from stop_words import safe_get_stop_words
+from stop_words import StopWordError
 from stop_words import STOP_WORDS_CACHE
+from stop_words import LANGUAGE_MAPPING
+from stop_words import AVAILABLE_LANGUAGES
 
 
 class StopWordsTestCase(TestCase):
@@ -29,7 +32,7 @@ class StopWordsTestCase(TestCase):
         sw = get_stop_words('fr')
         self.assertTrue('french' in STOP_WORDS_CACHE)
         original_stop_words_dir = stop_words.STOP_WORDS_DIR
-        stop_words.STOP_WORDS_DIR = '/trash/'
+        stop_words.STOP_WORDS_DIR = 'not-existing-directory'
         self.assertEqual(sw, get_stop_words('french'))
         stop_words.STOP_WORDS_DIR = original_stop_words_dir
         try:
@@ -43,7 +46,7 @@ class StopWordsTestCase(TestCase):
 
     def test_get_stop_words_install_issue(self):
         original_stop_words_dir = stop_words.STOP_WORDS_DIR
-        stop_words.STOP_WORDS_DIR = '/trash/'
+        stop_words.STOP_WORDS_DIR = 'not-existing-directory'
         self.assertRaises(StopWordError, get_stop_words, 'german')
         stop_words.STOP_WORDS_DIR = original_stop_words_dir
 
@@ -51,6 +54,13 @@ class StopWordsTestCase(TestCase):
         self.assertRaises(StopWordError, get_stop_words, 'huttese')
         self.assertEqual(safe_get_stop_words('huttese'), [])
 
+    def test_random_language_stop_words_load(self):
+        languages = list(LANGUAGE_MAPPING.keys()) + list(AVAILABLE_LANGUAGES)
+        sample = random.sample(languages, len(languages))
+        for language in sample:
+            stop_words = safe_get_stop_words(language)
+            self.assertTrue(len(stop_words) > 0,
+                            'Cannot load stopwords for %s language' % language)
 
 loader = TestLoader()